remove column_buffer_go18
